[test] Remove wiremock test server
- Fixes https://github.com/Azure/autorest.csharp/issues/1010
