--- test/AutoRest.TestServer.Tests/Infrastructure/TestServerVersion.cs
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace AutoRest.TestServer.Tests.Infrastructure
-{
-    public enum TestServerVersion
-    {
-        V1,
-        V2
-    }
-}
