+++ test/AutoRest.TestServer.Tests/NonStringEnumTests.cs
@@ -9,11 +9,8 @@ using NUnit.Framework;
 
 namespace AutoRest.TestServer.Tests
 {
-    [IgnoreOnTestServer(TestServerVersion.V2, "No record")]
     public class NonStringEnumTests : TestServerTestBase
     {
-        public NonStringEnumTests(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task NonStringEnumsGetFloat() => TestStatus(async (host, pipeline) =>
         {
