+++ test/AutoRest.TestServer.Tests/body-boolean.cs
@@ -11,8 +11,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class BooleanTests : TestServerTestBase
     {
-        public BooleanTests(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetBoolFalse() => Test(async (host, pipeline) =>
         {
