+++ test/AutoRest.TestServer.Tests/body-byte.cs
@@ -12,8 +12,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class BodyByte : TestServerTestBase
     {
-        public BodyByte(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetByteEmpty() => Test(async (host, pipeline) =>
         {
