+++ test/AutoRest.TestServer.Tests/body-date.cs
@@ -13,8 +13,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class BodyDateTest: TestServerTestBase
     {
-        public BodyDateTest(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetDateInvalid() => Test((host, pipeline) =>
         {
