+++ test/AutoRest.TestServer.Tests/body-dictionary.cs
@@ -16,8 +16,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class BodyDictionary : TestServerTestBase
     {
-        public BodyDictionary(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetDictionaryArrayEmpty() => Test(async (host, pipeline) =>
         {
@@ -559,7 +557,6 @@ namespace AutoRest.TestServer.Tests
         });
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Unmatched request")]
         public Task PutDictionaryDateTimeValid() => TestStatus(async (host, pipeline) =>
         {
             var value = new Dictionary<string, DateTimeOffset>
@@ -597,7 +594,6 @@ namespace AutoRest.TestServer.Tests
         });
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Unmatched request")]
         public Task PutDictionaryDoubleValid() => TestStatus(async (host, pipeline) =>
         {
             var value = new Dictionary<string, double>
@@ -628,7 +624,6 @@ namespace AutoRest.TestServer.Tests
         });
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Unmatched request")]
         public Task PutDictionaryFloatValid() => TestStatus(async (host, pipeline) =>
         {
             var value = new Dictionary<string, float>
