+++ test/AutoRest.TestServer.Tests/body-duration.cs
@@ -13,8 +13,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class BodyDurationTest: TestServerTestBase
     {
-        public BodyDurationTest(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetDurationInvalid() => Test((host, pipeline) =>
         {
