+++ test/AutoRest.TestServer.Tests/body-time.cs
@@ -11,11 +11,8 @@ using NUnit.Framework;
 
 namespace AutoRest.TestServer.Tests
 {
-    [IgnoreOnTestServer(TestServerVersion.V2, "No record")]
     public class BodyTimeTest: TestServerTestBase
     {
-        public BodyTimeTest(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task BodyTimeGet() => Test(async (host, pipeline) =>
         {
