+++ test/AutoRest.TestServer.Tests/model-flattening.cs
@@ -13,8 +13,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class ModelFlatteningTests : TestServerTestBase
     {
-        public ModelFlatteningTests(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task GetModelFlattenArray() => Test(async (host, pipeline) =>
         {
