+++ test/AutoRest.TestServer.Tests/object-type.cs
@@ -10,11 +10,8 @@ using object_type;
 
 namespace AutoRest.TestServer.Tests
 {
-    [IgnoreOnTestServer(TestServerVersion.V2, "Requests not matched")]
     public class ObjectTypeTest : TestServerTestBase
     {
-        public ObjectTypeTest(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task ObjectTypeErrorResponse() => Test((host, pipeline) =>
         {
