+++ test/AutoRest.TestServerLowLevel.Tests/url-query.cs
@@ -11,8 +11,6 @@ namespace AutoRest.TestServer.Tests
 {
     public class UrlQueryTests : TestServerLowLevelTestBase
     {
-        public UrlQueryTests(TestServerVersion version) : base(version) { }
-
         [Test]
         public Task UrlQueriesBoolTrue() => TestStatus(async (host) => await new QueriesClient(Key, host).GetBooleanTrueAsync());
 
@@ -32,22 +30,18 @@ namespace AutoRest.TestServer.Tests
         public Task UrlQueriesIntNull() => TestStatus(async (host) => await new QueriesClient(Key, host).GetIntNullAsync( null));
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Strict type checking, format from code model is incorrect")]
         public Task UrlQueriesLongPositive() => TestStatus(async (host) => await new QueriesClient(Key, host).GetTenBillionAsync());
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Strict type checking, format from code model is incorrect")]
         public Task UrlQueriesLongNegative() => TestStatus(async (host) => await new QueriesClient(Key, host).GetNegativeTenBillionAsync());
 
         [Test]
         public Task UrlQueriesLongNull() => TestStatus(async (host) => await new QueriesClient(Key, host).GetLongNullAsync( null));
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Recording match is too strict")]
         public Task UrlQueriesFloatPositive() => TestStatus(async (host) => await new QueriesClient(Key, host).FloatScientificPositiveAsync());
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Recording match is too strict")]
         public Task UrlQueriesFloatNegative() => TestStatus(async (host) => await new QueriesClient(Key, host).FloatScientificNegativeAsync());
 
         [Test]
@@ -63,7 +57,6 @@ namespace AutoRest.TestServer.Tests
         public Task UrlQueriesDoubleNull() => TestStatus(async (host) => await new QueriesClient(Key, host).DoubleNullAsync( null));
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Request not matched.")]
         public Task UrlQueriesStringUnicode() => TestStatus(async (host) => await new QueriesClient(Key, host).StringUnicodeAsync());
 
         [Test]
@@ -97,7 +90,6 @@ namespace AutoRest.TestServer.Tests
         public Task UrlQueriesDateNull() => TestStatus(async (host) => await new QueriesClient(Key, host).DateNullAsync( null));
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "Recording match is too strict")]
         public Task UrlQueriesDateTimeValid() => TestStatus(async (host) => await new QueriesClient(Key, host).DateTimeValidAsync());
 
         [Test]
@@ -107,7 +99,6 @@ namespace AutoRest.TestServer.Tests
         public Task UrlQueriesArrayCsvValid() => TestStatus(async (host) => await new QueriesClient(Key, host).ArrayStringCsvValidAsync( new[] {"ArrayQuery1", "begin!*'();:@ &=+$,/?#[]end", "", ""}));
 
         [Test]
-        [IgnoreOnTestServer(TestServerVersion.V2, "No recording")]
         public Task UrlQueriesArrayNoCollectionFormatValid() => TestStatus(async (host) => await new QueriesClient(Key, host).ArrayStringNoCollectionFormatEmptyAsync( new[] {"hello", "nihao", "bonjour"}));
 
         [Test]
